requirements.txt
psycopg2==2.9.10
SQLAlchemy==2.0.36
python-dotenv==1.0.1
